ability to wipe local storage
